[#124] 🐛 다크모드 수정
