+++ src/app/providers.tsx
@@ -1,8 +1,19 @@
 'use client'
-
 import { ThemeProvider } from 'next-themes'
+import { useEffect, useState } from 'react'
 
 export function Providers({ children }: { children: React.ReactNode }) {
+  const [mounted, setMounted] = useState(false)
+  useEffect(() => {
+    setMounted(true)
+    return () => {
+      setMounted(false)
+    }
+  }, [])
+
+  if (!mounted) {
+    return null
+  }
   return (
     <ThemeProvider attribute='class' defaultTheme='system' enableSystem>
       {children}
